release v2.1 - add tadl

+++ dubhe-server/common-k8s/src/main/java/org/dubhe/k8s/domain/vo/LogMonitoringVO.java
@@ -30,7 +30,7 @@ import java.util.List;
 @Data
 @AllArgsConstructor
 public class LogMonitoringVO extends PtBaseResult {
-    private Long totalLogs;
+    private Integer totalLogs;
     private List<String> logs;
 
     public LogMonitoringVO() {
